requirements.txt
 Django==1.11
 django-bootstrap3==8.2.2
 python-gitlab==0.20
